requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
ffmpeg-python==0.2.0
filelock==3.12.0
frozenlist==1.3.3
future==0.18.3
gTTS==2.3.2
hyperdb-python==0.1.3
idna==3.4
Jinja2==3.1.2
keyboard==0.13.5
llvmlite==0.40.0
MarkupSafe==2.1.2
more-itertools==9.1.0
mouse==0.7.1
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numba==0.57.0
numpy==1.24.3
openai==0.27.6
openai-whisper==20230314
PyAudio==0.2.13
pydub==0.25.1
python-dotenv==1.0.0
regex==2023.5.5
requests==2.30.0
sympy==1.11.1
tiktoken==0.3.1
torch==2.0.0
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==2.0.2
uuid==1.30
Wave==0.0.2
yarl==1.9.2
